+++ config/initializers/container.rb
@@ -46,6 +46,8 @@ module Charkeeper
     register('commands.characters_context.items.consume') { CharactersContext::Items::ConsumeCommand.new }
     register('commands.characters_context.change_feat') { CharactersContext::ChangeFeatCommand.new }
 
+    register('commands.characters_context.cosmere.create') { CharactersContext::Cosmere::CreateCommand.new }
+
     register('commands.characters_context.dc20.create') { CharactersContext::Dc20::CreateCommand.new }
     register('commands.characters_context.dc20.update') { CharactersContext::Dc20::UpdateCommand.new }